requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pandas
pydantic
python-multipart
scikit-learn
